test superuser permissions

+++ fuckme/app/routers/resources.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from fastapi import APIRouter
+from fastapi import APIRouter, status
 from fastapi.exceptions import HTTPException
 from fastapi.params import Depends, Query
 from sqlmodel import (
@@ -9,6 +9,7 @@ from sqlmodel import (
 )
 from sqlalchemy import func
 
+from app import deps
 from app.database import get_session
 from app.models import (
     Resource,
@@ -18,6 +19,7 @@ from app.models import (
     ResourceReadWithDetails,
     Topic,
     ResourceTopicLink,
+    User,
 )
 
 router = APIRouter(prefix="/resources", tags=["resources"])
@@ -32,9 +34,12 @@ def get_resource(session: Session, resource_id: int) -> Resource:
 
 @router.post("/", response_model=ResourceReadWithDetails)
 def create_resource(
-    *, session: Session = Depends(get_session), resource: ResourceCreate
+    *,
+    session: Session = Depends(get_session),
+    resource: ResourceCreate,
+    current_user: User = Depends(deps.get_current_user),
 ):
-    db_resource = Resource.from_orm(resource)
+    db_resource = Resource.from_orm(resource, {"user": current_user})
 
     # TODO(TOM): make this get or create topic
     db_resource.topics = [session.get(Topic, topic_id) for topic_id in resource.topics]
@@ -83,8 +88,11 @@ def update_resource(
     session: Session = Depends(get_session),
     resource_id: int,
     resource: ResourceUpdate,
+    current_user: User = Depends(deps.get_current_user),
 ):
     db_resource = get_resource(session, resource_id)
+    if not current_user.is_superuser and db_resource.user.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
     resource_data = resource.dict(exclude_unset=True)
     for key, value in resource_data.items():
         setattr(db_resource, key, value)
@@ -128,8 +136,15 @@ def mark_resource_broken(*, session: Session = Depends(get_session), resource_id
 
 
 @router.delete("/{resource_id}")
-def delete_resource(*, session: Session = Depends(get_session), resource_id: int):
+def delete_resource(
+    *,
+    session: Session = Depends(get_session),
+    resource_id: int,
+    current_user: User = Depends(deps.get_current_user),
+):
     resource = get_resource(session, resource_id)
+    if not current_user.is_superuser and resource.user.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
     session.delete(resource)
     session.commit()
     return {"ok": True}
